address CR suggestions
