+++ public/locales/fr/common.json
@@ -176,6 +176,7 @@
   "postalCode": "Code postal",
   "showPassword": "Afficher le mot de passe",
   "hidePassword": "Masquer le mot de passe",
+  "security": "Sécurité",
   "volunteer": "Bénévole",
   "moreCount": "+{{count}} de plus",
   "togglePledgedRaised": "Basculer entre les montants promis et collectés",
