+++ public/locales/fr/translation.json
@@ -1428,7 +1428,16 @@
     "enterState": "Entrez l'état",
     "select": "Sélectionner",
     "asYourCountry": "comme votre pays",
-    "actions": "Actions"
+    "actions": "Actions",
+    "changePassword": "Changer le mot de passe",
+    "oldPassword": "Ancien mot de passe",
+    "newPassword": "Nouveau mot de passe",
+    "password": "Mot de passe",
+    "updatePassword": "Mettre à jour le mot de passe",
+    "confirmNewPassword": "Confirmer le nouveau mot de passe",
+    "passCantBeEmpty": "Le mot de passe ne peut pas être vide",
+    "passNoMatch": "Le nouveau mot de passe et la confirmation ne correspondent pas",
+    "passwordChangedSuccessfully": "Mot de passe modifié avec succès"
   },
   "people": {
     "title": "Personnes",
