+++ public/locales/zh/common.json
@@ -50,6 +50,7 @@
   "desc": "描述",
   "enterPassword": "输入密码",
   "password": "密码",
+  "security": "安全",
   "confirmPassword": "确认密码",
   "forgotPassword": "忘记密码？",
   "talawaAdminPortal": "Talawa 管理门户",
