+++ public/locales/zh/translation.json
@@ -1428,7 +1428,16 @@
     "enterState": "输入州/省",
     "select": "请选择",
     "asYourCountry": "作为您的国家",
-    "actions": "操作"
+    "actions": "操作",
+    "changePassword": "更改密码",
+    "oldPassword": "旧密码",
+    "newPassword": "新密码",
+    "password": "密码",
+    "updatePassword": "更新密码",
+    "confirmNewPassword": "确认新密码",
+    "passCantBeEmpty": "密码不能为空",
+    "passNoMatch": "新密码和确认密码不一致",
+    "passwordChangedSuccessfully": "密码修改成功"
   },
   "userLogin": {
     "login": "登录",
