+++ src/screens/AdminPortal/MemberDetail/Security.module.css
@@ -1,0 +1,4 @@
+.securityCard {
+  border-radius: var(--radius-xl);
+  width: 100%;
+}
