+++ src/types/Auth/auth.ts
@@ -82,24 +82,3 @@ export interface IOAuthProviderConfig {
   redirectUri?: string;
   enabled: boolean;
 }
-
-export interface InterfacePasswordUpdateModalProps {
-  open: boolean;
-  onClose: () => void;
-  onSubmit: () => void;
-
-  values: {
-    oldPassword?: string;
-    newPassword: string;
-    confirmNewPassword: string;
-  };
-
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-
-  hidePreviousPassword?: boolean;
-  title: string;
-  saveText: string;
-  oldPasswordLabel: string;
-  newPasswordLabel: string;
-  confirmPasswordLabel: string;
-}
